Booking page fixed 2

+++ src/pages/Admin/AdminDashboard/admin-user/UserManagement.jsx
@@ -1,5 +1,6 @@
 import axios from 'axios'
 import React, { useEffect, useState } from 'react'
+import Swal from 'sweetalert2';
 
 const UserManagement = () => {
     const [users, setUsers] = useState([])
@@ -32,17 +33,45 @@ const UserManagement = () => {
 
     const handleDeleteUser = async (id) => {
         if (!id) return
-        try {
-            await axios.delete(`http://localhost:3000/api/users/${id}`, {
-                headers: {
-                    Authorization: `Bearer ${localStorage.getItem("token")}`
-                }
-            })
-            setUsers(users.filter(user => user._id !== id))
+        const result = await Swal.fire({
+            title: 'Are You Sure? You want to delete?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'Cancel',
+        })
+        if (result.isConfirmed) {
+            try {
+                await axios.delete(`http://localhost:3000/api/users/${id}`, {
+                    headers: {
+                        Authorization: `Bearer ${localStorage.getItem("token")}`
+                    }
+                })
+                setUsers(users.filter(user => user._id !== id))
+                Swal.fire({
+                    icon: 'success',
+                    title: 'User Deleted',
+                    text: 'The user has been deleted successfully.',
+                    showConfirmButton: false,
+                    timer: 1000,
+                })
 
-        } catch (error) {
-            console.error("Error in deleting user", error)
+
+            } catch (error) {
+                Swal.fire({
+                    icon: "error",
+                    title: "User Deletion failed",
+                    text: error.response?.data?.message || 'Error in deleting user. Please try again.'
+                })
+            }
         }
+
+
+
+
 
 
     }
